feat: Adicionando camanda de autorização de usuário aos endpoints PUT e DELETE

+++ fast_zero/app.py
@@ -1,10 +1,17 @@
-from fastapi import Depends, FastAPI, HTTPException
+from fastapi import Depends, FastAPI, HTTPException, status
+from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from fast_zero.models import User
 from fast_zero.database import get_session
-from fast_zero.schemas import UserPublic, UserSchema, UserList, Message
+from fast_zero.models import User
+from fast_zero.schemas import Message, UserList, UserPublic, UserSchema
+from fast_zero.security import (
+    create_access_token,
+    get_current_user,
+    get_password_hash,
+    verify_password,
+)
 
 app = FastAPI()
 
@@ -22,10 +29,14 @@ def create_user(user: UserSchema, session: Session = Depends(get_session)):
 
     if db_user:
         raise HTTPException(
-            status_code=400, detail='Username already registed'
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail='Username already registed',
         )
+
+    hashed_password = get_password_hash(user.password)
+
     db_user = User(
-        username=user.username, email=user.email, password=user.password
+        username=user.username, email=user.email, password=hashed_password
     )
     session.add(db_user)
     session.commit()
@@ -36,7 +47,7 @@ def create_user(user: UserSchema, session: Session = Depends(get_session)):
 
 @app.get('/users/', response_model=UserList)
 def read_users(
-    skip: int = 0, limit: int = 100, session: Session = Depends(get_session)
+    skip: int = 0, limit: int = 10, session: Session = Depends(get_session)
 ):
     users = session.scalars(select(User).offset(skip).limit(limit)).all()
     return {'users': users}
@@ -44,14 +55,26 @@ def read_users(
 
 @app.put('/users/{user_id}', response_model=UserPublic)
 def update_user(
-    user_id: int, user: UserSchema, session: Session = Depends(get_session)
+    user_id: int,
+    user: UserSchema,
+    session: Session = Depends(get_session),
+    current_user: User = Depends(get_current_user),
 ):
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail='Not enough permissions',
+        )
+
     db_user = session.scalar(select(User).where(User.id == user_id))
+
     if db_user is None:
-        raise HTTPException(status_code=404, detail='User not found')
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail='User not found'
+        )
     db_user.username = user.username
     db_user.email = user.email
-    db_user.password = user.password
+    db_user.password = get_password_hash(user.password)
     session.commit()
     session.refresh(db_user)
 
@@ -59,11 +82,45 @@ def update_user(
 
 
 @app.delete('/users/{user_id}', response_model=Message)
-def delete_user(user_id: int, session: Session = Depends(get_session)):
+def delete_user(
+    user_id: int,
+    session: Session = Depends(get_session),
+    current_user: User = Depends(get_current_user),
+):
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail='Not enough permissions',
+        )
     db_user = session.scalar(select(User).where(User.id == user_id))
     if db_user is None:
-        raise HTTPException(status_code=404, detail='User not found')
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail='User not found'
+        )
 
     session.delete(db_user)
     session.commit()
     return {'detail': 'User deleted'}
+
+
+@app.post('/token')
+def login_for_access_token(
+    form_data: OAuth2PasswordRequestForm = Depends(),
+    session: Session = Depends(get_session),
+):
+    user = session.scalar(select(User).where(User.email == form_data.username))
+
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail='Incorrect email or password',
+        )
+
+    if not verify_password(form_data.password, user.password):
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail='Incorrect email or password',
+        )
+
+    access_token = create_access_token(data={'sub': user.email})
+    return {'access_token': access_token, 'type_token': 'bearer'}
